requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
google-generativeai>=0.8.0
python-dotenv==1.0.0
Pillow==10.2.0
scikit-image==0.21.0
opencv-python==4.8.0.74
joblib==1.3.2
scikit-learn==1.3.2
scipy==1.11.4

